% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{build.cpt}
\alias{build.cpt}
\title{Build cpts for a node given a description of the parents}
\usage{
build.cpt(df)
}
\arguments{
\item{df}{a dataframe with rows for each parent node and columns
as described in notes below}
}
\value{
a dataframe with columns for each of the parent nodes and the
two output probabilities, and a row for each possible combination of
the parent values. The positive output probability is named with
the value in the "positive" column of the dataframe.
}
\description{
Constructs a dataframe with columns describing all possible values
of the parent nodes and probabilities that their child node will
have a certain categorical value. Only works for computing nodes
with two possible categorical values.
}
\note{
the dataframe df must have a row for each parent node and columns
"parent" containing the name of the parent node, func containing the name
of the function used to compute the weights as described in
\code{\link{get.favorable.weight}}. If the function is not "general"
it must also have columns for the base, weight, and hilow as described
in \code{\link{get.favorable.weight}}.
}

