% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cnet.test}
\alias{cnet.test}
\title{More complex test network}
\format{\preformatted{Formal class 'catNetwork' [package "catnet"] with 14 slots
  ..@ objectName     : chr "catNetwork"
  ..@ numnodes       : int 7
  ..@ nodes          : chr [1:7] "a" "s" "p" "q" ...
  ..@ meta           : chr ""
  ..@ maxParents     : int 3
  ..@ parents        :List of 7
  .. ..$ : NULL
  .. ..$ : NULL
  .. ..$ : int 2
  .. ..$ : int [1:3] 2 3 6
  .. ..$ : int 4
  .. ..$ : NULL
  .. ..$ : int 5
  ..@ maxCategories  : int 2
  ..@ categories     :List of 7
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  .. ..$ : chr [1:2] "1" "2"
  ..@ probabilities  :List of 7
  .. ..$ : num [1:2] 0.97 0.03
  .. ..$ : num [1:2] 0.387 0.613
  .. ..$ :List of 2
  .. .. ..$ : num [1:2] 0.405 0.595
  .. .. ..$ : num [1:2] 0.619 0.381
  .. ..$ :List of 2
  .. .. ..$ :List of 2
  .. .. .. ..$ :List of 2
  .. .. .. .. ..$ : num [1:2] 0.708 0.292
  .. .. .. .. ..$ : num [1:2] 0.374 0.626
  .. .. .. ..$ :List of 2
  .. .. .. .. ..$ : num [1:2] 0.637 0.363
  .. .. .. .. ..$ : num [1:2] 0.95 0.05
  .. .. ..$ :List of 2
  .. .. .. ..$ :List of 2
  .. .. .. .. ..$ : num [1:2] 0.357 0.643
  .. .. .. .. ..$ : num [1:2] 0.179 0.821
  .. .. .. ..$ :List of 2
  .. .. .. .. ..$ : num [1:2] 0.619 0.381
  .. .. .. .. ..$ : num [1:2] 0.458 0.542
  .. ..$ :List of 2
  .. .. ..$ : num [1:2] 0.361 0.639
  .. .. ..$ : num [1:2] 0.199 0.801
  .. ..$ : num [1:2] 0.252 0.748
  .. ..$ :List of 2
  .. .. ..$ : num [1:2] 0.357 0.643
  .. .. ..$ : num [1:2] 0.369 0.631
  ..@ complexity     : int 17
  ..@ nodeComplexity : num [1:7] 1 1 2 8 2 1 2
  ..@ likelihood     : num 0
  ..@ nodeLikelihood : logi NA
  ..@ nodeSampleSizes: logi NA
}}
\usage{
cnet.test
}
\description{
A sample \code{\link{catNetwork}} used to test functions. The
same as the "cnet2" network used in the catnet vignette
}
\keyword{datasets}

