% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.r
\name{dash20s}
\alias{dash20s}
\title{Turn dashes to zeros}
\usage{
dash20s(x)
}
\arguments{
\item{x}{a character vector to be cleaned}
}
\description{
Take a character vector of numbers where dashes indicate negligible
values, and return a numeric vector where dashes have been set to 0.
}

