% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix.marginals}
\alias{fix.marginals}
\title{Fix the marginal probabilities of a node}
\usage{
fix.marginals(d, n, p)
}
\arguments{
\item{d}{a \code{\link{catNetwork}} model object}

\item{n}{the number of the node to modify}

\item{p}{the desired new marginal probabilities. It must be a numeric
vector with values between 0 and 1, summing to 1, with as many elements
as there are category levels of node n.}
}
\value{
a \code{\link{catNetwork}} model object identical to d, except the marginal
probabilities of node n have been altered.
}
\description{
Set the marginal probabilities of a node in the model by adjusting
the underlying conditional probabilities in the catNetwork object.
}

