% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.r
\name{gcd}
\alias{gcd}
\title{Greatest common denominator}
\usage{
gcd(x, y)
}
\arguments{
\item{x}{the first integer. Must be the larger of the two.}

\item{y}{the second integer.}
}
\description{
Find the greatest common denominator of two integers
}

