% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.margins}
\alias{generate.margins}
\title{Generate a range of marginal probabilities}
\usage{
generate.margins(n = 2)
}
\arguments{
\item{n}{the number of columns to generate. Defaults to 2.}
}
\value{
a dataframe with n columns
}
\description{
Generate a dataframe with a varying number of columns. Values are
constrained to range from 0.1 to 0.9 in increments of 0.1, and row
totals must sum to 1
}

