% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genie2catnet}
\alias{genie2catnet}
\title{Load a GeNiE model from an .xdsl file and return a catnet object}
\usage{
genie2catnet(fn)
}
\arguments{
\item{fn}{the filename to read in}
}
\value{
a \code{\link{catNetwork}} object representation of the GeNiE network
given by the filename
}
\description{
Given a filename this function loads it as XML, extracts the nodes,
cpt's, and parents and returns a \code{\link{catNetwork}} object.
}

