% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{get.favorable.weight}
\alias{get.favorable.weight}
\title{Calculate a set of parent node weights}
\usage{
get.favorable.weight(df, levels, base = 0.1, minw = 0.1, maxw = 0.9)
}
\arguments{
\item{df}{a dataframe with rows for each parent node and columns
as described in notes below}

\item{levels}{a vector of observed states for the parent nodes}

\item{base}{the base (starting) probability of the positive outcome}

\item{minw}{the minimum probability of the positive outcome. If set to
NULL no minimum is checked. Defaults to 0.1.}

\item{maxw}{the maximum probability of the postive outcome. If set to
NULL no maximum is checked. Defaults to 0.9.}
}
\value{
a probability for the node to have the "Favorable" value
as determined by labels used outside of this function.
}
\description{
Calculates a set of output state probabilities given a dataframe
with rows for each parent of a node and columns
describing how to calculate the weights for each parent, a set of
observed values for each parent, and a base probability of the
positive (mathematically) outcome.
}
\note{
the dataframe df must have a row for each parent node and columns
parent containing the name of the parent node, func containing the name
of the function used to compute the weights. "func" can be "general"
meaning the generic linear model approach is used, or the name of the
weighting function to be used (retrieved from the local
environment with the \code{\link{get}} function)
which is then passed the df dataframe and
levels vector.

The weight column is a vector of weights for the parent nodes when
using the general generic linear approach.

There are also columns "hilow1", "hilow2", and "hilow3"
listing possible values of the parent node in order from most positive
to least positive effect on the output.

The parent node values in levels must be in the same order as the rows
of df.
}

