% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{get.relative.weight}
\alias{get.relative.weight}
\title{Get the relative weight of a parent node's value}
\usage{
get.relative.weight(obs, values, wt = NULL)
}
\arguments{
\item{obs}{the observed level of the parent node.}

\item{values}{a vector of possible values the parent node might have,
in order from most positive to most negative.}

\item{weight}{a vector of weights to assign each possible value. If
NULL it will assign weights of 1, 0 to values with two levels,
or 1, 0, -1 to values with three levels. For more than three levels
weights needs to be explicitly assigned.}
}
\value{
a relative weight for a parent node value. Using default
parameters either 1, 0, or -1.
}
\description{
Calculate the relative weight of a parent node on a child node.
Takes a vector of possible values of the parent node and returns
a scaled weighting on the child node, usually in (0, 1) or (-1, 0, 1).
}

