% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPtable}
\alias{getPtable}
\title{Extract a node's probability table}
\usage{
getPtable(m, n, d = NULL)
}
\arguments{
\item{m}{a catnet model}

\item{n}{a node name}

\item{d}{a single row dataframe with column names for nodes in
the model and fixed values for the state of each node}
}
\description{
Extract the probability table from a node in a catnet model
with cnProb and return it as a dataframe.
}

