% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marginals.R
\name{margin.loglik}
\alias{margin.loglik}
\title{Generate the likelihood profile for a node's margins}
\usage{
margin.loglik(d, nname, data)
}
\arguments{
\item{d}{a \code{\link{catNetwork}} model object}

\item{nname}{the numeric node for a model OR a character name for the node}

\item{data}{a dataframe of sample data for the model, in either character
names of levels or numbers for the category level of each node}
}
\value{
a dataframe with the columns for the marginal probabilities for each
possible value of the node, and a column for the log-likelihood for those
margins given the data.
}
\description{
Given a model and a sample of data from that model, calculate the
log-likelihood for all possible combinations of marginal probabilities
for a node.
}
\note{
"All possible combinations" is an exageration, it's actually limited
to the range 0.1 - 0.9 in increments of 0.1 as generated by generate.margins().
}

