% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ncat}
\alias{ncat}
\title{Possibly output a status message}
\usage{
ncat(noisy, trigger = 0, ...)
}
\arguments{
\item{noisy}{a flag, set to differing values depending upon how many
output you want}

\item{trigger}{if the flag is greater than this value call
cat(). Defautls to zero.}
}
\description{
If the value passed is greater than the threshold, pass the remaining
arguments to cat() to be sent to the output stream. Useful for tracking
progress of a long script with tuneable feedback.
}

