% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ntimestamp}
\alias{ntimestamp}
\title{Possibly output a timestamp}
\usage{
ntimestamp(noisy, trigger = 0)
}
\arguments{
\item{noisy}{a flag, set to differing values depending upon how many
timestamps you want}

\item{trigger}{if the flag is greater than this value call
timestamp(). Defautls to zero.}
}
\description{
If the value passed is greater than the threshold, call the timestamp()
function.
}

