% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parent.parms}
\alias{parent.parms}
\title{Find a node's parents}
\usage{
parent.parms(nid, nodes, parlist, catlist)
}
\arguments{
\item{nid}{node of interest}

\item{nodes}{vector of node names}

\item{parlist}{list of vectors of parents for each node}

\item{catlist}{list of vectors of categories for each node}
}
\description{
This function returns a dataframe  containing a node's parents (in order)
and the corresponding number of levels
}

