% Generated by roxygen2 (4.0.1): do not edit by hand
\name{populate.genie}
\alias{populate.genie}
\title{Populate a GeNiE model with species data}
\usage{
populate.genie(fn, spp, fn.out = fn, NA.syn = c("none"), nodelist = NULL)
}
\arguments{
\item{fn}{filename of the GeNiE .xdsl file}

\item{spp}{a dataframe corresponding to rows of species and columns
of nodes in the GeNiE model, identifying the value of each node. The
dataframe must have a column called "sp_name" to identify each species.}

\item{fn.out}{output filename -- defaults to the input file name}

\item{NA.syn}{synonyms for NA's or missign values. Defaults to c("none")}

\item{nodelist}{a character vector of node names to attempt to populate.
Defaults to all nodes deterumined by column names. "sp_name" and
an optional column named "sp_code" are ignored.}
}
\value{
No value returned, but creates a new .xdsl file containing the
newly populated GeNiE model
}
\description{
Given a GeNiE model file this function will attempt to populate it
with information in a dataframe with rows corresponding to species
and columns corresponding to nodes in the model with an appropriate
value (or NA) in each cell
}

