% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{q2levels}
\alias{q2levels}
\title{Convert a continuous vector to discrete values}
\usage{
q2levels(x, q = 0.5, lv = c("Low", "High"), absolute = TRUE,
  eps = 1e-04)
}
\arguments{
\item{x}{vector of numeric values}

\item{q}{a single cutpoint for the quantile function. Defaults to 0.5.}

\item{lv}{a character vector of (2) levels to name the categories. Defaults
c("Low", "High")}

\item{absolute}{if TRUE, then the cutpoint given in q is taken as
given. If FALSE they it is used as the cutpoint for vector x. Defaults
to TRUE}

\item{eps}{the small amount to adjust the cutpoint in the case of
overrun as described above. Defaults to 0.0001}
}
\description{
Take a vector of numeric values and return a vector of categorical values
based on either quantiles or absolute values. If there are multiple
duplicates at the extremes so that a cutoff is "overrun" (e.g., 50% zeros
with a lower cutpoint of 0.33) the cutpoint is set to be just outside the
minimum or maximum value.
}

