% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{q3levels}
\alias{q3levels}
\title{Convert a continuous vector to discrete values}
\usage{
q3levels(x, q = c(0.33, 0.66), lv = c("Low", "Medium", "High"),
  absolute = TRUE, eps = 1e-04)
}
\arguments{
\item{x}{vector of numeric values}

\item{q}{vector of (2) cutpoints for the quantile function. Defaults to
c(0.33, 0.66)}

\item{lv}{a character vector of (3) levels to name the categories. Defaults
c("Low", "Medium", "High")}

\item{absolute}{if TRUE, then the cutpoints in q are taken as
given. If FALSE they are used as as cutpoints for quantiles of vector x.
Defaults to TRUE.}

\item{eps}{the small amount to adjust the cutpoints in the case of
overrun as described above. Defaults to 0.0001}
}
\description{
Take a vector of numeric values and return a vector of categorical values
based on either quantiles or absolute values. If there are multiple
duplicates at the extremes so that a cutoff is "overrun" (e.g., 50% zeros
with a lower cutpoint of 0.33) the cutpoint is set to be just outside the
minimum or maximum value.
}

