% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataprocessing.R
\name{q5levels}
\alias{q5levels}
\title{Convert a continuous vector to discrete values}
\usage{
q5levels(x, q = c(0.2, 0.4, 0.6, 0.8), lv = c("Very_small", "Small",
  "Medium", "Large", "Very_large"), absolute = TRUE, eps = 1e-04)
}
\arguments{
\item{x}{vector of numeric values}

\item{q}{vector of (4) cutpoints for the quantile function. Defaults to 20%
in each category}

\item{lv}{a character vector of (5) levels to name the categories. Defaults
range from "Very_small" to "Very_large".}

\item{absolute}{if TRUE, then the cutpoints in q are taken as
given. If FALSE they are used as as cutpoints for quantiles of vector x.
Defaults to TRUE.}

\item{eps}{the small amount to adjust the cutpoints in the case of
overrun as described above. Defaults to 0.0001}
}
\description{
Take a vector of numeric values and return a vector of categorical values
based on either quantiles or absolute values. If there are multiple
duplicates at the extremes so that a cutoff is "overrun" (e.g., 50% zeros
with a lower cutpoint of 0.33) the cutpoint is set to be just outside the
minimum or maximum value.
}

