% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{set.cpt}
\alias{set.cpt}
\title{Set the cpt for a node in a catNet model}
\usage{
set.cpt(N, nname, cptdf)
}
\arguments{
\item{N}{a catNet model}

\item{nname}{the name of a node to alter in model N}

\item{cptdf}{a dataframe consisting of rows for all possible combinations
of node nname's parents, and probabilities for node nname's value in each
combination. More details below.}
}
\value{
a catNet model the same as N, with the probabilities for node
nname updated to match those in cptdf.
}
\description{
Uses a dataframe of parent node value combinations and value probabilities
(for instance, as generated by \code{\link{build.cpt}})
to set the cpt for a node in a catNet model.
}
\note{
the dataframe cptdf must have column names for each of the parent
nodes, plus two columns for the probabilities of node states for each
combination. The second column from the right (column ncol(cptdf) - 1)
must have a name matching one of the defined values for node nname.
}

