% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{set.node.levels}
\alias{set.node.levels}
\title{Set or get a vector of node values}
\usage{
set.node.levels(N, nname, value = NULL)
}
\arguments{
\item{N}{a catNet model}

\item{nname}{the name of a node to alter in model N}

\item{value}{the desired value of the node. Stops on an error if the
value is not one specified in the catNet model. If NULL it returns a
vector of legal values.}
}
\value{
a named vector, whose name is the specified nodename (nname). If
value is NULL the vector contains all legal values.
}
\description{
Uses a dataframe of parent node value combinations and value probabilities
(for instance, as generated by \code{\link{build.cpt}})
to set the cpt for a node in a catNet model.
}

