% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{simpnet}
\alias{simpnet}
\title{Simple test network}
\format{\preformatted{Formal class 'catNetwork' [package "catnet"] with 14 slots
  ..@ objectName     : chr "catNetwork"
  ..@ numnodes       : int 3
  ..@ nodes          : chr [1:3] "a" "b" "c"
  ..@ meta           : chr ""
  ..@ maxParents     : int 2
  ..@ parents        :List of 3
  .. ..$ : NULL
  .. ..$ : int 1
  .. ..$ : int [1:2] 1 2
  ..@ maxCategories  : int 2
  ..@ categories     :List of 3
  .. ..$ : chr [1:2] "a1" "a2"
  .. ..$ : chr [1:2] "b1" "b2"
  .. ..$ : chr [1:2] "c1" "c2"
  ..@ probabilities  :List of 3
  .. ..$ : num [1:2] 0.2 0.8
  .. ..$ :List of 2
  .. .. ..$ : num [1:2] 0.6 0.4
  .. .. ..$ : num [1:2] 0.4 0.6
  .. ..$ :List of 2
  .. .. ..$ :List of 2
  .. .. .. ..$ : num [1:2] 0.3 0.7
  .. .. .. ..$ : num [1:2] 0.7 0.3
  .. .. ..$ :List of 2
  .. .. .. ..$ : num [1:2] 0.9 0.1
  .. .. .. ..$ : num [1:2] 0.1 0.9
  ..@ complexity     : int 7
  ..@ nodeComplexity : num [1:3] 1 2 4
  ..@ likelihood     : num 0
  ..@ nodeLikelihood : logi NA
  ..@ nodeSampleSizes: logi NA
}}
\usage{
simpnet
}
\description{
A simple \code{\link{catNetwork}} used to test functions.
}
\keyword{datasets}

