% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model weighting.R
\name{summarize.output.nodes}
\alias{summarize.output.nodes}
\title{Summarize the posterior output nodes of a model}
\usage{
summarize.output.nodes(N, data, targets)
}
\arguments{
\item{N}{a catNet model}

\item{data}{a dataframe of data for the model. Rows represent
observations and columns are nodes in the model. NA values in a
column mean the node is to be estimated.}

\item{targets}{a character vector of node names to be summarized}
}
\value{
a dataframe with columns for the node name, the value of
the node, and the proportion of results with that value.
}
\description{
Given a model and a set of data, calculate the distribution of
values for a set of unknown, modeled nodes.
}

