% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catnetparsing.R
\name{which.ordered}
\alias{which.ordered}
\title{Find which elements in a vector match those of another}
\usage{
which.ordered(x, v)
}
\arguments{
\item{x}{a vector of values to be found in vector v}

\item{v}{a vector of values to search in}
}
\value{
a vector of indexes of v, indicating which elements match those
        in x, in the same order they occur in x
}
\description{
This function finds which elements of vector v match those in
vector x and return matching indexes of v in the same order they
appear in x.
}
\note{
Which returns matched indexes in increasing order, and
network model nodes need to be parsed in the the proper order relative
to the model structure.
}
\keyword{ordered}
\keyword{which}

