% Generated by roxygen2 (4.0.1): do not edit by hand
\name{conditionalP}
\alias{conditionalP}
\title{Find the conditional probability of a node}
\usage{
conditionalP(m, n, d = NULL)
}
\arguments{
\item{m}{a \code{\link{catNetwork}} object}

\item{n}{the node of interest}

\item{d}{a single-row dataframe with column names equal to the model
node names and values indicating the state of each node}
}
\description{
Return the conditional probability of a node from
a cnet object, given a model, node, and data row.
}

