% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catnetparsing.R
\name{getPtable}
\alias{getPtable}
\title{Extract a node's probability table}
\usage{
getPtable(m, n, d = NULL)
}
\arguments{
\item{m}{a catnet model}

\item{n}{a node name}

\item{d}{a single row dataframe with column names for nodes in
the model and fixed values for the state of each node}
}
\value{
a dataframe with rows for all possible combinations of
the node's parents and marginal probabilities of node values
for each combination.
}
\description{
Extract the probability table from a node in a catnet model
with cnProb and return it as a dataframe.
}

