% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggpng}
\alias{ggpng}
\title{Write out a graphic to a png file}
\usage{
ggpng(P, fn, x = 6.5, y = 6.5 * 3/4, uni = "in", rez = 300,
  gdir = NULL)
}
\arguments{
\item{object}{to plot (usually generated by ggplot)}

\item{fn}{file name (without .png suffix)}

\item{x}{horizontal dimensions of default graphic (defaults to 6.5)}

\item{y}{vertical dimensinos of default graphic (defaults to 4.875)}

\item{uni}{units of measurement for the dimension (defaults to "in")}

\item{rez}{resolution of output graphic (pixels per unit)}
}
\description{
Write a ggplot object to a .png file. Defaults to 300dpi resolution on
"letter" sized paper in portrait orientation with standard margins.
Automatically adds a ".png" suffix to the provided filename.
}
\seealso{
\code{\link{png}}, \code{\link{pdf}}
}

