% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.r
\name{sizeof}
\alias{sizeof}
\title{display the size of an R object in appropirate units}
\usage{
sizeof(x)
}
\arguments{
\item{x}{an object in the environment}
}
\description{
Display the size of an R object in scale-appropriate units. Sometimes
useful for deciding if a particular object should be re-generated as
needed or saved and loaded on repeat analysis.
}

